Replacing libraries with vzwtab equivalents
Three libraries used the __android_log_print symbol but did not list
liblog.so as a dependency.  This was causing unresolved symbol errors
when android HAL was dlopen()'ing the libraries.  Apparently it's very
difficult to add liblog.so as DT_NEEDED in the .dynamic section of the
library, so I grabbed the equivalent libraries from a vzwtab.  The
libraries on the vzwtab do not call __android_log_print.